requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
idna==3.4
psycopg2==2.9.5
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
aiohttp==3.8.4
aioresponses==0.7.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
builtwith==1.3.4
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-environ==0.10.0
frozenlist==1.3.3
httpretty==1.1.4
idna==3.4
lxml==4.9.2
multidict==6.0.4
psycopg2==2.9.5
python-Wappalyzer==0.3.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
yarl==1.8.2
